extracts db interactions into labeled files
