+++ Dockerfile
@@ -45,7 +45,6 @@ ENV NODE_ENV=production
 ENV NEXT_TELEMETRY_DISABLED=1
 ENV NEXT_RUNTIME="nodejs"
 
-RUN npm run migrate
 RUN addgroup --system --gid 1001 nodejs
 RUN adduser --system --uid 1001 nextjs
