+++ src/db/interactions/offers.ts
@@ -1,0 +1,22 @@
+import { offers } from "@/db/schema";
+import { db } from "@/db";
+
+export const createOffer = async (offer: any, vehicleId = 0) => {
+  if (!offer?.code || !offer?.validUntil) return;
+  try {
+    return await db
+      .insert(offers)
+      .values({
+        ...offer,
+        vehicleId,
+        validUntil: new Date(offer?.validUntil),
+      })
+      .returning();
+  } catch (e) {
+    console.log("error creating offer record", e);
+  }
+};
+
+export const getAllOffers = async () => {
+  return await db.select().from(offers);
+};
